Implemented the view function to render volonteer list
